split STT result in 2 components: modal and result itself so we can display the results without the modal in the dedicated result page
