+++ src/components/AssistantPlainTextResult.vue
@@ -1,0 +1,197 @@
+<template>
+	<div class="plain-text-result">
+		<span class="assistant-bubble">
+			<CreationIcon :size="16" class="icon" />
+			<span>{{ t('assistant', 'Nextcloud Assistant') }}</span>
+		</span>
+		<h2>
+			{{ taskCategoryName }}
+		</h2>
+		<div v-if="myOutput"
+			class="output">
+			<label
+				for="assistant-output"
+				class="input-label">
+				{{ t('assistant', 'Output') }}
+			</label>
+			<NcRichContenteditable
+				id="assistant-output"
+				ref="output"
+				:value.sync="myOutput"
+				class="editable-output"
+				:multiline="true"
+				:placeholder="t('assistant', 'Output')"
+				:link-autocomplete="false" />
+			<NcNoteCard
+				class="warning-note"
+				type="warning">
+				{{ t('assistant', 'This output was generated by AI. Make sure to double-check and adjust.') }}
+			</NcNoteCard>
+			<div class="button-wrapper">
+				<NcButton type="secondary"
+					:title="t('assistant', 'Copy output text to clipboard')"
+					@click="onCopy">
+					{{ t('assistant', 'Copy output') }}
+					<template #icon>
+						<ClipboardCheckOutlineIcon v-if="copied" />
+						<ContentCopyIcon v-else />
+					</template>
+				</NcButton>
+				<NcButton :disabled="output === myOutput"
+					type="secondary"
+					:title="t('assistant', 'Reset the output value to the originally generated one')"
+					@click="delayedReset">
+					{{ t('assistant', 'Reset') }}
+				</NcButton>
+			</div>
+		</div>
+	</div>
+</template>
+
+<script>
+import CreationIcon from 'vue-material-design-icons/Creation.vue'
+import ClipboardCheckOutlineIcon from 'vue-material-design-icons/ClipboardCheckOutline.vue'
+import ContentCopyIcon from 'vue-material-design-icons/ContentCopy.vue'
+
+import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
+import NcRichContenteditable from '@nextcloud/vue/dist/Components/NcRichContenteditable.js'
+import NcNoteCard from '@nextcloud/vue/dist/Components/NcNoteCard.js'
+
+import VueClipboard from 'vue-clipboard2'
+import Vue from 'vue'
+
+import { showError } from '@nextcloud/dialogs'
+
+import { TASK_TYPES } from '../constants.js'
+
+Vue.use(VueClipboard)
+
+export default {
+	name: 'AssistantPlainTextResult',
+	components: {
+		NcButton,
+		NcRichContenteditable,
+		NcNoteCard,
+		CreationIcon,
+		ClipboardCheckOutlineIcon,
+		ContentCopyIcon,
+	},
+	props: {
+		output: {
+			type: String,
+			default: '',
+		},
+		taskCategory: {
+			type: [Number, null],
+			default: null,
+		},
+	},
+	data() {
+		return {
+			myOutput: this.output,
+			copied: false,
+		}
+	},
+	computed: {
+		taskCategoryName() {
+			switch (this.taskCategory) {
+			case TASK_TYPES.text_generation:
+				return t('assistant', 'Text Generation')
+			case TASK_TYPES.speech_to_text:
+				return t('assistant', 'Speech to Text')
+			default:
+				return t('assistant', 'Unknown Result Type')
+			}
+		},
+	},
+	mounted() {
+		if (this.myOutput !== '') {
+			this.$refs.output.focus()
+		}
+	},
+	methods: {
+		async onCopy() {
+			try {
+				const container = this.$refs.output.$el
+				await this.$copyText(this.output.trim(), container)
+				this.copied = true
+				setTimeout(() => {
+					this.copied = false
+				}, 5000)
+			} catch (error) {
+				console.error(error)
+				showError(t('assistant', 'Result could not be copied to clipboard'))
+			}
+		},
+		onReset() {
+			this.myOutput = this.output
+		},
+
+		delayedReset() {
+			// This is a hack to ensure the text box is updated
+			// when we reset the text since removing newlines or spaces
+			// from the end of the text does not trigger an update.
+
+			// Delete any trailing newlines
+			this.myOutput = this.myOutput.replace(/\n+$/, '')
+			this.myOutput += '.'
+
+			// Let the ui refresh before resetting the text
+			setTimeout(() => {
+				this.onReset()
+			}, 0)
+		},
+	},
+}
+</script>
+
+<style lang="scss" scoped>
+.plain-text-result {
+	//width: 100%;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	overflow-y: auto;
+
+	> h2 {
+		display: flex;
+		margin: 12px 0 20px 0;
+		.icon {
+			margin-right: 8px;
+		}
+	}
+
+	.editable-output {
+		width: 100%;
+		display: flex;
+		flex-direction: column;
+	}
+
+	.assistant-bubble {
+		align-self: start;
+		display: flex;
+		gap: 8px;
+		background-color: var(--color-primary-element-light);
+		border-radius: var(--border-radius-rounded);
+		padding: 2px 8px;
+		.icon {
+			color: var(--color-primary);
+		}
+	}
+
+	.button-wrapper {
+		display: flex;
+		flex-direction: row;
+		align-items: center;
+		justify-content: center;
+		margin-top: 12px;
+		margin-bottom: 12px;
+
+		>* {
+			margin-right: 12px;
+			margin-left: 12px;
+		}
+	}
+}
+</style>
