+++ src/views/PlainTextResultPage.vue
@@ -1,8 +1,10 @@
 <template>
 	<NcContent app-name="assistant">
 		<NcAppContent>
-			<div v-if="task.output">
-				<AssistantPlainTextModal
+			<div v-if="task.output"
+				class="assistant-wrapper">
+				<AssistantPlainTextResult
+					class="form"
 					:output="task.output"
 					:task-category="task.category" />
 			</div>
@@ -14,7 +16,7 @@
 import NcContent from '@nextcloud/vue/dist/Components/NcContent.js'
 import NcAppContent from '@nextcloud/vue/dist/Components/NcAppContent.js'
 
-import AssistantPlainTextModal from '../components/AssistantPlainTextModal.vue'
+import AssistantPlainTextResult from '../components/AssistantPlainTextResult.vue'
 
 import { showError } from '@nextcloud/dialogs'
 
@@ -22,9 +24,9 @@ export default {
 	name: 'PlainTextResultPage',
 
 	components: {
+		AssistantPlainTextResult,
 		NcContent,
 		NcAppContent,
-		AssistantPlainTextModal,
 	},
 
 	props: {
