+++ templates/speechToTextResultPage.php
@@ -5,7 +5,3 @@ use OCP\Util;
 
 $appId = Application::APP_ID;
 Util::addScript($appId, $appId . '-speechToTextResultPage');
-
-?>
-
-<div id="assistant-stt-content"></div>
